requirements.txt
asgiref==3.5.2
Django==4.1.3
django-crispy-forms==1.14.0
sqlparse==0.4.3
django-import-export==3.0.1